pyproject.toml
[tool.poetry]
name = "auto-encoder-other"
version = "0.1.0"
description = ""
authors = ["wilfred11 <wilfredvancasteren@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.10"
torch = "^2.8.0"
torchvision = "^0.23.0"
pandas = "^2.3.2"
matplotlib = "^3.10.5"
seaborn = "^0.13.2"
numpy = "^2.3.2"
scikit-learn = "^1.7.1"
tqdm = "^4.67.1"
celluloid = "^0.2.0"
ffmpeg-python = "^0.2.0"
pathlib = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
